requirements.txt
speechrecognition==3.10.4
pydub==0.25.1
faiss-cpu==1.8.0
ollama==0.1.8
gradio==4.31.0
groq==0.30.0
python-dotenv==1.0.0
pymupdf==1.24.10
langchain==0.1.16
langchain-community==0.0.33
